# backend/openapi.yml
openapi: 3.0.0
info:
  title: Memory Game API
  version: 1.0.0
  description: Backend API for the Memory Game application

servers:
  - url: http://localhost:4000
    description: Local development

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 6
          example: secret123

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: john_doe
        password:
          type: string
          example: secret123

    UserResponse:
      type: object
      properties:
        username:
          type: string
          example: john_doe

    MessageResponse:
      type: object
      properties:
        message:
          type: string

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful, confirmation email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Registration successful. Please check your email to confirm your account.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/verify/{token}:
    get:
      summary: Verify email address
      tags: [Auth]
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: Email verification token from the confirmation email
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Email verified successfully. You can now log in.
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      summary: Log in and receive an HttpOnly session cookie
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, HttpOnly cookie set
          headers:
            Set-Cookie:
              description: HttpOnly JWT session cookie (token=...; HttpOnly; SameSite=Strict)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Email not confirmed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/me:
    get:
      summary: Get the currently authenticated user
      tags: [Auth]
      responses:
        '200':
          description: Authenticated user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated or session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/logout:
    post:
      summary: Log out and clear the session cookie
      tags: [Auth]
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Logged out

  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok